# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: MM API
  version: 0.1.0
  description: |
    LINE/LIFF 前提の家計管理 API（初期スコープ）。
    本仕様は Expense CRUD と月次一覧取得を対象とする。
servers:
  - url: http://localhost:8787
    description: Local development
tags:
  - name: Expenses
    description: Expense resource operations
  - name: Settlements
    description: Monthly settlement operations
security:
  - bearerAuth: []
paths:
  /api/expenses:
    get:
      tags:
        - Expenses
      summary: 月次の支払い一覧を取得
      operationId: listExpensesByMonth
      parameters:
        - name: date
          in: query
          required: true
          description: 対象年月（YYYY-MM）
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: '2026-02'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListExpensesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Expenses
      summary: 支払いを作成
      operationId: createExpense
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/expenses/{id}:
    get:
      tags:
        - Expenses
      summary: 支払い詳細を取得
      operationId: getExpense
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Expenses
      summary: 支払いを更新
      operationId: updateExpense
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpenseRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Expenses
      summary: 支払いを削除
      operationId: deleteExpense
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: 削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/expenses/{id}/restore:
    post:
      tags:
        - Expenses
      summary: 削除した支払いを復元
      operationId: restoreExpense
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 復元成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/settlements:
    get:
      tags:
        - Settlements
      summary: 月次精算サマリを取得
      operationId: getSettlementByMonth
      parameters:
        - name: month
          in: query
          required: true
          description: 対象年月（YYYY-MM）
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: '2026-06'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementSummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ExpenseId:
      name: id
      in: path
      required: true
      description: Expense ID
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Retry-safe mutation key used to deduplicate retried create, update, and delete requests.
      schema:
        type: string
        minLength: 1
        maxLength: 255
  responses:
    BadRequest:
      description: リクエスト不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Invalid request
            details:
              field: date
              reason: must be YYYY-MM
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            required:
              summary: 認証情報なし
              value:
                message: 認証情報が見つかりません
                details:
                  code: AUTH_REQUIRED
                  action: LINEから開き直して、もう一度お試しください
            invalid:
              summary: 認証情報が無効
              value:
                message: 認証情報の有効期限が切れているか、正しくありません
                details:
                  code: AUTH_INVALID
                  action: LINEから開き直して、もう一度お試しください
    Forbidden:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: このLINEユーザーは家計簿に登録されていません
            details:
              code: USER_NOT_REGISTERED
              action: 初回登録で使うユーザーを選択してください
    ServiceUnavailable:
      description: 認証基盤または外部連携の一時エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: ユーザー情報を確認できません
            details:
              code: AUTH_UNAVAILABLE
              action: 家計簿のユーザー情報を読み込めません。時間をおいて再度お試しください。解消しない場合は管理者に連絡してください
    NotFound:
      description: 対象なし
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Expense not found
            details:
              id: "exp_123"
    Conflict:
      description: 更新競合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Expense version conflict
            details:
              id: "exp_123"
              expectedVersion: 3
              actualVersion: 4
    InternalServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Internal server error
            details: {}
  schemas:
    Expense:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - date
        - price
        - category
        - version
      properties:
        id:
          type: string
          example: exp_123
        userId:
          type: string
          example: user_1
        userName:
          type: string
          example: げんき
        date:
          type: string
          format: date
          example: '2026-02-15'
        price:
          type: integer
          minimum: 0
          example: 1200
        category:
          type: string
          example: food
        memo:
          type: string
          nullable: true
          example: lunch
        version:
          type: integer
          minimum: 1
          example: 1
    CreateExpenseRequest:
      type: object
      additionalProperties: false
      required:
        - date
        - price
        - category
      properties:
        date:
          type: string
          format: date
          example: '2026-02-15'
        price:
          type: integer
          minimum: 0
          example: 1200
        category:
          type: string
          example: food
        memo:
          type: string
          nullable: true
          example: lunch
    UpdateExpenseRequest:
      type: object
      additionalProperties: false
      required:
        - version
      properties:
        date:
          type: string
          format: date
          example: '2026-02-15'
        price:
          type: integer
          minimum: 0
          example: 1200
        category:
          type: string
          example: food
        memo:
          type: string
          nullable: true
          example: lunch
        version:
          type: integer
          minimum: 1
          example: 1
      minProperties: 2
    ListExpensesResponse:
      type: object
      additionalProperties: false
      required:
        - expenses
      properties:
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
    ExpenseDetailResponse:
      type: object
      additionalProperties: false
      required:
        - expense
        - deleted
      properties:
        expense:
          $ref: '#/components/schemas/Expense'
        deleted:
          type: boolean
          example: false
    SettlementSummaryResponse:
      type: object
      additionalProperties: false
      required:
        - month
        - householdTotal
        - userTotals
        - difference
        - settlement
      properties:
        month:
          type: string
          pattern: '^\d{4}-\d{2}$'
          example: '2026-06'
        householdTotal:
          type: integer
          minimum: 0
          example: 14000
        userTotals:
          type: array
          items:
            $ref: '#/components/schemas/UserMonthlyTotal'
        difference:
          type: integer
          minimum: 0
          example: 6000
        settlement:
          $ref: '#/components/schemas/MonthlySettlement'
    UserMonthlyTotal:
      type: object
      additionalProperties: false
      required:
        - userId
        - displayName
        - total
      properties:
        userId:
          type: string
          example: user_1
        displayName:
          type: string
          example: A
        total:
          type: integer
          minimum: 0
          example: 10000
    MonthlySettlement:
      type: object
      additionalProperties: false
      required:
        - fromUserId
        - toUserId
        - amount
      properties:
        fromUserId:
          type: string
          nullable: true
          example: user_2
        toUserId:
          type: string
          nullable: true
          example: user_1
        amount:
          type: integer
          minimum: 0
          example: 3000
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - message
        - details
      properties:
        message:
          type: string
          example: Invalid request
        details:
          type: object
          additionalProperties: true